tests/wpt/web-platform-tests/trusted-types/script-src.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<body>
<script>
  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.safe);

    var s = document.createElement('script');
    s.src = url;
    assert_equals(url + '', s.src);
  });

  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.javascript);

    var s = document.createElement('script');
    s.src = url;
    assert_equals(url + '', s.src);
  });

  test(t => {
    var url = TrustedScriptURL.unsafelyCreate(URLS.external);

    var s = document.createElement('script');
    s.src = url;
    assert_equals(url + '', s.src);
  });
</script>
